# docs/openapi.yaml
openapi: 3.0.3
info:
  title: STG - API de persistencia
  description: 'Contrato formal de la capa de persistencia del ecosistema Smart Booking STG.

    Generado a partir de las rutas de Express (src/routes/*.ts).


    Cubre el hallazgo C-4 de la auditoria: los endpoints estaban documentados en la

    Tabla 7 de la tesis en prosa, sin contrato verificable que permitiera probarlos.


    Recuento verificado en esta version: 73 rutas sobre 16 recursos.

    '
  version: 1.0.0
servers:
- url: http://localhost:3001
  description: Entorno local de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido por POST /api/admin-users/login
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Clave fija para los flujos de n8n (requerida si AUTH_ENABLED=true)
security:
- bearerAuth: []
- apiKeyAuth: []
paths:
  /api/admin-users:
    get:
      tags:
      - adminUsers
      summary: GET /api/admin-users
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
    post:
      tags:
      - adminUsers
      summary: POST /api/admin-users
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
        '201':
          description: Creado
...
  /api/clientes:
    get:
      tags:
      - clientes
      summary: GET /api/clientes
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
    post:
      tags:
      - clientes
      summary: POST /api/clientes
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
        '201':
          description: Creado
  /api/clientes/{id}:
    get:
      tags:
      - clientes
      summary: GET /api/clientes/{id}
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    put:
      tags:
      - clientes
      summary: PUT /api/clientes/{id}
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      tags:
      - clientes
      summary: DELETE /api/clientes/{id}
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/cotizaciones:
    get:
      tags:
      - cotizaciones
      summary: GET /api/cotizaciones
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
    post:
      tags:
      - cotizaciones
      summary: POST /api/cotizaciones
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
        '201':
          description: Creado
...
  /api/destinos:
    get:
      tags:
      - destinos
      summary: GET /api/destinos
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
    post:
      tags:
      - destinos
      summary: POST /api/destinos
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
        '201':
          description: Creado
...
  /api/documentos:
    get:
      tags:
      - documentos
      summary: GET /api/documentos
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
...
  /api/estadisticas:
    get:
      tags:
      - estadisticas
      summary: GET /api/estadisticas
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
...
  /api/hoteles:
    get:
      tags:
      - hoteles
      summary: GET /api/hoteles
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
    post:
      tags:
      - hoteles
      summary: POST /api/hoteles
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
        '201':
          description: Creado
...
  /api/pagos:
    get:
      tags:
      - pagos
      summary: GET /api/pagos
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
    post:
      tags:
      - pagos
      summary: POST /api/pagos
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
        '201':
          description: Creado
...
  /api/parametros:
    get:
      tags:
      - parametros
      summary: GET /api/parametros
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
...
  /api/pasajeros:
    get:
      tags:
      - pasajeros
      summary: GET /api/pasajeros
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
    post:
      tags:
      - pasajeros
      summary: POST /api/pasajeros
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
        '201':
          description: Creado
...
  /api/recordatorios:
    get:
      tags:
      - recordatorios
      summary: GET /api/recordatorios
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
    post:
      tags:
      - recordatorios
      summary: POST /api/recordatorios
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
        '201':
          description: Creado
...
  /api/reservas:
    get:
      tags:
      - reservas
      summary: GET /api/reservas
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
    post:
      tags:
      - reservas
      summary: POST /api/reservas
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
        '201':
          description: Creado
...
  /api/tramos:
    get:
      tags:
      - tramos
      summary: GET /api/tramos
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
    post:
      tags:
      - tramos
      summary: POST /api/tramos
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
        '201':
          description: Creado
...
  /api/viajes:
    get:
      tags:
      - viajes
      summary: GET /api/viajes
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
    post:
      tags:
      - viajes
      summary: POST /api/viajes
      responses:
        '200':
          description: OK
        '400':
          description: Datos invalidos (validacion Zod)
        '401':
          description: No autenticado (si AUTH_ENABLED=true)
        '404':
          description: No encontrado
        '201':
          description: Creado
...